requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
audioop-lts==0.2.1
blinker==1.9.0
click==8.1.8
discord==2.3.2
discord.py==2.5.2
ffmpeg==1.4
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.3.0
python-dotenv==1.0.1
typing_extensions==4.12.2
wavelink==3.4.1
Werkzeug==3.1.3
yarl==1.18.3
yt-dlp==2025.2.19
